unit test recurrent NetX load/save
